+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.3.0
 - You can now scan 2d matrix codes (FMD) and it will automaticly add the product to the durability list with durability date.
+- Double clicking the products will open a new window where you can change location and date.
 
 ## v0.2.0
 - You can now add custom products to the database.